feat: Refactor Face Scanner for improved modularity
This commit refactors the Face Scanner feature to improve its structure, readability, and maintainability.

The key changes include:
- Breaking down the monolithic `facescantemp.vue` component into smaller, more manageable components: `FaceScanner.vue`, `FaceScannerCamera.vue`, and `RegistrationModal.vue`.
- Abstracting the `face-api.js` logic into a dedicated `faceapi.service.ts` to decouple face recognition from the UI.
- Moving the authentication logic into an `auth.face.service.ts` to separate concerns.
- Cleaning up the codebase by removing unused variables and commented-out code.
- Updating the router to use the new `FaceScanner.vue` component.
