+++ .gitignore
@@ -36,3 +36,4 @@ package-lock.json
 package.json
 
 resources/android/**/*
+public/models/
