+++ src/components/face-scanner/ClockInModal.vue
